refactor(web): move the relative-time formatter beside the shared lib
The contract list and detail views need the same coarse relative rendering
the notification rows use, and a second copy in the contracts feature would
be two definitions of one format. A contracts module importing it from the
notifications feature would state a dependency that does not exist, so it
moves to src/lib beside useDocumentTitle instead.

Pure move: the function, its tests, and the one call site are unchanged
apart from the import path.
